requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-WTF==1.2.1